refactor(market): PR #341 코멘트 반영
- ConflictInfo, StockWithConflict를 domain/types.ts로 이동 (Architecture #1)
- conflict-types.ts 삭제
- resolveConflicts를 utils/resolveConflicts.ts로 추출 (CONVENTIONS.md)
- SignalStockCard props를 StockWithConflict 직접 사용 (TypeScript #5)
- SignalSubsection.stocks 타입을 StockWithConflict[]로 정밀화
- ⓘ 버튼에 title, aria-label 추가
