--- src/components/dashboard/conflict-types.ts
@@ -1,10 +1,0 @@
-import type { StockSignalResult } from '@/domain/types';
-
-export interface ConflictInfo {
-    readonly bullishCount: number;
-    readonly bearishCount: number;
-}
-
-export type StockWithConflict = StockSignalResult & {
-    readonly conflict?: ConflictInfo;
-};
